rewrote search function as component in header, rather than its own page, since it's better for ux.  Also changed limits on search results to match newer limit numbers.
