+++ client/pages/index.js
@@ -33,15 +33,12 @@ const Home = () => {
     rootMargin: '-100px 0px',
   };
 
-  const [refSearchOptionsTitle, inViewSearchOptionsTitle] = useInView(
-    useInViewOptions
-  );
-  const [refSearchOptionsBody, inViewSearchOptionsBody] = useInView(
-    useInViewOptions
-  );
-  const [refMemberBenefitsTitle, inViewMemberBenefitsTitle] = useInView(
-    useInViewOptions
-  );
+  const [refSearchOptionsTitle, inViewSearchOptionsTitle] =
+    useInView(useInViewOptions);
+  const [refSearchOptionsBody, inViewSearchOptionsBody] =
+    useInView(useInViewOptions);
+  const [refMemberBenefitsTitle, inViewMemberBenefitsTitle] =
+    useInView(useInViewOptions);
   const [refMemberBenefitsBody, inViewMemberBenefitsBody] = useInView(
     useInViewOptionsMemberItems
   );
@@ -226,7 +223,7 @@ const Home = () => {
             viewBox='0 0 1440 320'
           >
             <path
-              fill='#55BA82'
+              fill='#10B981'
               fillOpacity='1'
               d='M0,192L48,186.7C96,181,192,171,288,149.3C384,128,480,96,576,122.7C672,149,768,235,864,240C960,245,1056,171,1152,122.7C1248,75,1344,53,1392,42.7L1440,32L1440,320L1392,320C1344,320,1248,320,1152,320C1056,320,960,320,864,320C768,320,672,320,576,320C480,320,384,320,288,320C192,320,96,320,48,320L0,320Z'
             ></path>
