+++ client/pages/search/results.js
@@ -16,7 +16,7 @@ const SearchResultsPage = ({ drinks }) => {
   const [page, setPage] = useState(1);
 
   const [order, setOrder] = useState('asc');
-  const [limit, setLimit] = useState(10);
+  const [limit, setLimit] = useState(25);
 
   const queryString = router.query.search;
